backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
click==8.1.8
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.1
setuptools==76.0.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
uv==0.6.10
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
sqlalchemy
pydantic_settings
psycopg2
python-jose
redis
bcrypt==4.0.1
alembic
boto3
pgvector
mistralai
PyPDF2
celery
tiktoken
langchain-openai
langchain